//Translação (em 2D e em 3D) //[DONE]○ translate2D(vector, dx, dy) //[DONE]○ translate3D(vector, dx, dy, dz)
//Rotação (em 2D e em todos os eixos em 3D)
//[DONE] roration2D(vector, angle)
//[DONE] rotation3DX(vector, angle)
//[DONE] rotation3DY(vector, angle)
//[DONE] rotation3DZ(vector, angle)

//● Reflexão (em todos os eixos nas dimensões 2D e 3D)
//[DONE] reflection2DX(vector)
//[DONE] reflection2DY(vector)
//[DONE] reflection3DX(vector)
//[DONE] reflection3DY(vector)
//[DONE] reflection3DZ(vector)

//● Projeção (em todos os eixos em 2D e m 3D)
//[DONE] projection2DX(vector)
//[DONE] projection2DY(vector)
//[DONE] projection3DX(vector)
//[DONE] projection3DY(vector)
//[DONE] projection3DZ(vector)

//● Cisalhamento (apenas em 2D)
//[DONE] shearing(vector, kx, ky

+++ src/main/java/av2/Vector.java
@@ -2,11 +2,11 @@ package av2;
 
 public class Vector {
 
-    private int vector[];
+    private double vector[];
     private int dimension;
 
-    public Vector(int dim, int[] elements) {
-        vector = new int[dim];
+    public Vector(int dim, double[] elements) {
+        vector = new double[dim];
         this.dimension = dim;
 
         for (int i = 0; i < dim; i++) {
@@ -15,12 +15,12 @@ public class Vector {
         }
     }
 
-    public int get(int index) {
+    public double get(int index) {
 
         return vector[index];
     }
 
-    public void set(int index, int element) {
+    public void set(int index, double element) {
 
         vector[index] = element;
